+++ test/models/test_mnist.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 import unittest
 import numpy as np
+from tinygrad.state import get_parameters
 from tinygrad.tensor import Tensor, Device
 from tinygrad.nn import optim, BatchNorm2d
 from extra.training import train, evaluate
@@ -16,7 +17,7 @@ class TinyBobNet:
     self.l2 = Tensor.scaled_uniform(128, 10)
 
   def parameters(self):
-    return optim.get_parameters(self)
+    return get_parameters(self)
 
   def forward(self, x):
     return x.dot(self.l1).relu().dot(self.l2).log_softmax()
@@ -38,7 +39,7 @@ class TinyConvNet:
       self.bn1, self.bn2 = lambda x: x, lambda x: x
 
   def parameters(self):
-    return optim.get_parameters(self)
+    return get_parameters(self)
 
   def forward(self, x:Tensor):
     x = x.reshape(shape=(-1, 1, 28, 28)) # hacks
